refactor: migrate to TS, settup eslint && add dependencies
