+++ .eslintrc.json
@@ -1,4 +1,8 @@
 {
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "project": "./tsconfig.json"
+  },
   "env": {
     "browser": true,
     "es2021": true,
@@ -6,27 +10,23 @@
     "jest": true
   },
   "extends": [
+    "plugin:react/recommended",
+    "plugin:@typescript-eslint/eslint-recommended",
     "airbnb-base",
-    "plugin:prettier/recommended",
+    "airbnb-typescript/base",
     "eslint:recommended",
-    "prettier",
-    "plugin:react/recommended"
+    "plugin:prettier/recommended",
+    "prettier"
   ],
-  "plugins": ["prettier", "import", "react"],
-  "parserOptions": {
-    "parser": "@babel/eslint-parser",
-    "requireConfigFile": false,
-    "ecmaVersion": 2018,
-    "jsx": true,
-    "sourceType": "module"
-  },
+  "plugins": ["prettier", "import", "react", "@typescript-eslint", "react-hooks"],
   "rules": {
+    "react-hooks/rules-of-hooks": "error",
+    "react-hooks/exhaustive-deps": "warn",
     "prettier/prettier": "warn",
     "no-debugger": "off",
     "no-console": "warn",
     "class-methods-use-this": "off",
     "import/no-dynamic-require": "off",
-    // "max-lines-per-function": ["error", { "max": 60, "skipBlankLines": false, "skipComments": true }],
     "import/prefer-default-export": "off",
     "import/no-unresolved": "off",
     "react/prop-types": "off",
@@ -34,5 +34,11 @@
     "react/react-in-jsx-scope": 0,
     "react/jsx-uses-react": 0,
     "no-unused-vars": "warn"
+  },
+  "settings": {
+    "react": {
+      "pragma": "React",
+      "version": "detect"
+    }
   }
 }
